+++ Sources/SwiftKit/iCloud/URL+iCloud.swift
@@ -11,7 +11,8 @@ import Foundation
 public extension URL {
     
     /**
-     The url to the iCloud ubiquity container.
+     The url to the iCloud ubiquity container. This is where
+     documents and data will be saved and synced with iCloud.
      */
     static func ubiquityContainer(
         for manager: FileManager = .default,
@@ -20,12 +21,22 @@ public extension URL {
     }
     
     /**
-     The url to the iCloud ubiquity container Documents root.
+     The url to the iCloud ubiquity container Documents root,
+     where documents will be saved and synced with iCloud.
      */
     static func ubiquityContainerDocuments(
         for manager: FileManager = .default,
         containerId: String? = nil) -> URL? {
         ubiquityContainer(for: manager, containerId: containerId)?
             .appendingPathComponent("Documents")
+    }
+    
+    /**
+     The url to a local document fallback directory that can
+     be used when the `ubiquityContainer` urls are nil.
+     */
+    static func ubiquityContainerDocumentsLocalFallbackDirectory(
+        for manager: FileManager = .default) -> URL? {
+        manager.urls(for: .documentDirectory, in: .userDomainMask).first
     }
 }