+++ Release Notes.md
@@ -10,7 +10,7 @@ I will bump revision by revision, until SwiftKit has all functionality that it s
 * `iCloudDocumentSync` is a new protocol for syncing iCloud document changes. 
 * `String+Slugify` is a new extension that can convert a string to a slugified version.
 * `StandardiCloudDocumentSync` is a new class for syncing iCloud document changes.
-* `URL+iCloud` contains iCloud-specific URL extensions.
+* `URL+iCloud` contains iCloud-specific URLs and fallback URLs.